Split ppo indicator into several files.
