+++ js/masters/indicators/indicators-all.src.js
@@ -35,7 +35,7 @@ import '../../Stock/Indicators/MFI/MFIIndicator.js';
 import '../../Stock/Indicators/Momentum/MomentumIndicator.js';
 import '../../Stock/Indicators/NATR/NATRIndicator.js';
 import '../../Stock/Indicators/PivotPointsIndicator.js';
-import '../../Stock/Indicators/PPOIndicator.js';
+import '../../Stock/Indicators/PPO/PPOIndicator.js';
 import '../../Stock/Indicators/PC/PCIndicator.js';
 import '../../Stock/Indicators/PriceEnvelopesIndicator.js';
 import '../../Stock/Indicators/PSAR/PSARIndicator.js';
