+++ ts/Stock/Indicators/PPO/PPOPoint.d.ts
@@ -1,0 +1,33 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+import PPOIndicator from './PPOIndicator';
+import type EMAPoint from '../EMA/EMAPoint';
+
+/* *
+*
+*  Class
+*
+* */
+
+declare class PPOPoint extends EMAPoint {
+    public series: PPOIndicator;
+}
+
+/* *
+*
+*  Default Export
+*
+* */
+
+export default PPOPoint;
